# Sawagger/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
...
    "/admin/login": {
      "post": {
        "description": "使用用户名密码登录，返回 JWT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "管理员管理"
        ],
        "summary": "管理员登录",
        "parameters": [
          {
            "description": "用户名和密码",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AdminService.AdminLoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/adminSwaggerResponse.AdminLoginSuccessResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/adminSwaggerResponse.AdminLoginErrorResponse"
            }
          }
        }
      }
    },
...
    "/user/change_pwd": {
      "put": {
        "description": "已登录用户通过旧密码修改为新密码",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户"
        ],
        "summary": "修改密码",
        "parameters": [
          {
            "description": "修改密码请求体",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserService.ChangePasswordSchema"
            }
          },
          {
            "type": "string",
            "description": "用户 JWT Token",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "密码修改成功",
            "schema": {
              "$ref": "#/definitions/userSwaggerResponse.ChangePasswordForgetSuccessResponse"
            }
          },
          "400": {
            "description": "参数错误或修改失败",
            "schema": {
              "$ref": "#/definitions/userSwaggerResponse.ChangePasswordErrorResponse"
            }
          }
        }
      }
    },
    "/user/forget": {
      "put": {
        "description": "用户忘记密码时，通过用户名和邀请码重置密码",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户"
        ],
        "summary": "忘记密码",
        "parameters": [
          {
            "description": "忘记密码请求体",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserService.ForgetUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "密码重置成功",
            "schema": {
              "$ref": "#/definitions/userSwaggerResponse.ForgetSuccessResponse"
            }
          },
          "400": {
            "description": "参数错误或重置失败",
            "schema": {
              "$ref": "#/definitions/userSwaggerResponse.ForgetErrorResponse"
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "description": "使用用户名和密码登录，返回不同状态码：\n- code = -1：普通错误（返回累计错误次数）\n- code = -2：达到锁定阈值（返回剩余解锁秒数，用于前端倒计时）\n- code = -3：异常请求（例如请求过于频繁，疑似机器人）\n- code = 1：登录成功（返回 token）",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户"
        ],
        "summary": "用户登录",
        "parameters": [
          {
            "description": "登录请求参数",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserService.UserLogin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "登录成功,返回token",
            "schema": {
              "$ref": "#/definitions/userSwaggerResponse.LoginSuccessResponse"
            }
          },
          "400": {
            "description": "登录失败，-1普通错误，-2限制登录倒计时，-3异常请求",
            "schema": {
              "$ref": "#/definitions/userSwaggerResponse.LoginErrorResponse"
            }
          }
        }
      }
    },
    "/user/register": {
      "post": {
        "description": "使用用户名、密码和邀请码注册用户",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户"
        ],
        "summary": "注册新用户",
        "parameters": [
          {
            "description": "用户注册信息",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserService.CreateUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "注册成功",
            "schema": {
              "$ref": "#/definitions/swaggerResponse.SuccessResponseDoc"
            }
          },
          "400": {
            "description": "参数错误或注册失败",
            "schema": {
              "$ref": "#/definitions/userSwaggerResponse.RegisterUserErrorResponse"
            }
          }
        }
      }
    },
    "/user/rename": {
      "put": {
        "description": "修改当前登录用户的用户名",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户"
        ],
        "summary": "修改用户名",
        "parameters": [
          {
            "type": "string",
            "description": "新的用户名",
            "name": "username",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "用户 JWT Token",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "修改成功",
            "schema": {
              "$ref": "#/definitions/userSwaggerResponse.RenameUserSuccessResponse"
            }
          },
          "400": {
            "description": "参数错误 / 用户名已存在 / 用户不存在",
            "schema": {
              "$ref": "#/definitions/userSwaggerResponse.RenameUserErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "AdminService.AdminLoginRequest": {
      "type": "object",
      "required": [
        "auth_status",
        "device_id",
        "ip",
        "password",
        "username"
      ],
      "properties": {
        "auth_status": {
          "description": "记住登录状态：1-记住 2-不记住",
          "type": "integer"
        },
        "device_id": {
          "description": "设备唯一标识，用于多设备管理",
          "type": "string"
        },
        "ip": {
          "description": "当前登录IP地址",
          "type": "string"
        },
        "password": {
          "description": "登录密码，必填",
          "type": "string"
        },
        "username": {
          "description": "登录用户名，必填",
          "type": "string"
        }
      }
    },
...
    "UserService.ChangePasswordSchema": {
      "type": "object",
      "required": [
        "new_password",
        "old_password"
      ],
      "properties": {
        "new_password": {
          "type": "string"
        },
        "old_password": {
          "type": "string"
        }
      }
    },
    "UserService.CreateUser": {
      "type": "object",
      "required": [
        "code",
        "password",
        "username"
      ],
      "properties": {
        "code": {
          "type": "string"
        },
        "contact_info": {
          "description": "联系信息-用于找回密码",
          "type": "string"
        },
        "contact_type": {
          "description": "联系渠道-用于找回密码",
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "UserService.ForgetUser": {
      "type": "object",
      "required": [
        "code",
        "password",
        "username"
      ],
      "properties": {
        "code": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "UserService.UserLogin": {
      "type": "object",
      "required": [
        "password",
        "username"
      ],
      "properties": {
        "password": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
...
    "adminSwaggerResponse.AdminLoginErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": -1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "密码错误"
        }
      }
    },
    "adminSwaggerResponse.AdminLoginSuccessResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1N..."
        },
        "msg": {
          "type": "string",
          "example": "ok"
        }
      }
    },
...
    "swaggerResponse.SuccessResponseDoc": {
      "type": "object",
      "properties": {
        "code": {
          "description": "状态码",
          "type": "integer",
          "example": 1
        },
        "data": {
          "description": "数据",
          "type": "string",
          "example": "any"
        },
        "msg": {
          "description": "信息",
          "type": "string",
          "example": "ok"
        }
      }
    },
    "userSwaggerResponse.ChangePasswordErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": -1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "err msg"
        }
      }
    },
    "userSwaggerResponse.ChangePasswordForgetSuccessResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "ok"
        }
      }
    },
    "userSwaggerResponse.ForgetErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": -1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "err msg"
        }
      }
    },
    "userSwaggerResponse.ForgetSuccessResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "ok"
        }
      }
    },
    "userSwaggerResponse.LoginErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "example": "-1/-2/-3"
        },
        "data": {
          "type": "string",
          "example": "-1:null;-2:{num: 3, seconds: 60}; -3:null"
        },
        "msg": {
          "type": "string",
          "example": "err msg"
        }
      }
    },
    "userSwaggerResponse.LoginSuccessResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "string",
          "example": "{token: eyJhbGciOiJIUzI1N...}"
        },
        "msg": {
          "type": "string",
          "example": "ok"
        }
      }
    },
    "userSwaggerResponse.RegisterUserErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": -1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "err msg"
        }
      }
    },
    "userSwaggerResponse.RenameUserErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": -1
        },
        "data": {
          "type": "string",
          "example": "null"
        },
        "msg": {
          "type": "string",
          "example": "参数错误 / 用户名已存在 / 用户不存在"
        }
      }
    },
    "userSwaggerResponse.RenameUserSuccessResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 1
        },
        "data": {
          "type": "string",
          "example": "修改成功"
        },
        "msg": {
          "type": "string",
          "example": "ok"
        }
      }
    }
  }
}